feat: 1. 修改 items 对应信息获取 2. 修改利润等数据类型，某些物品价格很高超过 i32 范围了 3. 增加官方商品过滤逻辑 4. 增加武器等物品堆叠利润计算逻辑 5. 删除 tools-cli 项目

+++ rust/crates/nodex/src/profit_panel.rs
@@ -219,7 +219,7 @@ impl ProfitPanel {
 #[class(init,base=Control)]
 pub struct ProfitItem {
     pub label: String,
-    pub single_value: i32,
+    pub single_value: i64,
     pub total_value: i64,
     pub percentage: f32,
     #[base]
@@ -254,7 +254,7 @@ impl ProfitItem {
     pub fn set_value(
         &mut self,
         label: String,
-        single_value: i32,
+        single_value: i64,
         total_value: i64,
         percentage: f32,
     ) {
